refactor redaction into SRP modules
